\name{writeWorkbook}
\alias{writeWorkbook}
\title{Write a list of data frames into an Excel workbook}
\usage{
  writeWorkbook(L, filename, ...)
}
\arguments{
  \item{L}{The list of data frames to write.}

  \item{filename}{The name of the output file.}

  \item{...}{Additional arguments to
  \code{\link{writeWorksheet}}}
}
\description{
  Write a list of data frames into an Excel workbook
}

